Add support for local manifests
