+++ ern-core/src/GitManifest.ts
@@ -4,21 +4,17 @@ import { gitCli } from './gitCli'
 import _ from 'lodash'
 import fs from 'fs-extra'
 import path from 'path'
-import semver from 'semver'
 import Platform from './Platform'
 import log from './log'
 import kax from './kax'
+import { LocalManifest } from './LocalManifest'
 
-const manifestFileName = 'manifest.json'
-const pluginConfigFileName = 'config.json'
-const ERN_VERSION_DIRECTORY_RE = /ern_v(.+)\+/
-
-export default class GitManifest {
+export default class GitManifest extends LocalManifest {
   public readonly remote?: string
   public readonly branch?: string
   private readonly git: any
 
-  private cachedManifest: any
+  private hasSynced: boolean
 
   public constructor(
     public readonly repoLocalPath: string,
@@ -28,11 +24,11 @@ export default class GitManifest {
       remote: 'origin',
     }
   ) {
-    fs.ensureDirSync(repoLocalPath)
+    super(repoLocalPath)
     this.git = gitCli(repoLocalPath)
     this.remote = remote
     this.branch = branch
-    this.cachedManifest = null
+    this.hasSynced = false
   }
 
   public async sync() {
@@ -75,11 +71,6 @@ export default class GitManifest {
       await this.git.reset(['--hard', `${this.remote}/master`])
       await this.git.pull(this.remote, this.branch)
     }
-
-    const pathToManifestJson = path.join(this.repoLocalPath, manifestFileName)
-    this.cachedManifest = JSON.parse(
-      fs.readFileSync(pathToManifestJson, 'utf-8')
-    )
   }
 
   /**
@@ -91,21 +82,22 @@ export default class GitManifest {
    *  you might want to improve the code to act a bit smarter.
    */
   public async syncIfNeeded() {
-    if (!this.cachedManifest) {
+    if (!this.hasSynced) {
       await kax
         .task(`Syncing ${this.repoRemotePath || this.repoLocalPath} Manifest`)
         .run(this.sync())
+      this.hasSynced = true
     }
   }
 
   public async getManifest(): Promise<any> {
     await this.syncIfNeeded()
-    return this.cachedManifest
+    return super.getManifest()
   }
 
   public async hasManifestId(manifestId: string): Promise<boolean> {
-    const manifest = await this.getManifest()
-    return !Array.isArray(manifest) && manifest[manifestId]
+    await this.syncIfNeeded()
+    return super.hasManifestId(manifestId)
   }
 
   public async getManifestData({
@@ -115,217 +107,15 @@ export default class GitManifest {
     manifestId?: string
     platformVersion?: string
   } = {}): Promise<any | void> {
-    const manifest = await this.getManifest()
-
-    return Array.isArray(manifest)
-      ? // Old manifest format [to be deprecated]
-        _.find(manifest, m =>
-          semver.satisfies(platformVersion, m.platformVersion)
-        )
-      : // New manifest format
-        manifest[manifestId]
+    await this.syncIfNeeded()
+    return super.getManifestData({ manifestId, platformVersion })
   }
 
-  /**
-   * Return an array containing all top level plugin configuration directories
-   * are matching an Electrode Native version lower or equal than the specified
-   * maxVersion.
-   *
-   * For example, given the following directories :
-   *
-   *  /Users/blemair/.ern/ern-override-manifest/plugins/ern_v0.5.0+
-   *  /Users/blemair/.ern/ern-override-manifest/plugins/ern_v0.10.0+
-   *  /Users/blemair/.ern/ern-override-manifest/plugins/ern_v0.13.0+
-   *
-   * And maxVersion='0.10.0'
-   *
-   * The function would return :
-   *
-   *  [
-   *   '/Users/blemair/.ern/ern-override-manifest/plugins/ern_v0.5.0+',
-   *   '/Users/blemair/.ern/ern-override-manifest/plugins/ern_v0.10.0+'
-   *  ]
-   * @param maxVersion The upper bound Electrode Native version.
-   *                   Only plugin configuration directories lower than this
-   *                   version will be returned.
-   */
-  public getPluginsConfigurationDirectories(
-    maxVersion: string = Platform.currentVersion
-  ): string[] {
-    const pathToPluginsDirectory = path.join(this.repoLocalPath, 'plugins')
-    return fs.existsSync(pathToPluginsDirectory)
-      ? _(fs.readdirSync(pathToPluginsDirectory))
-          .filter(
-            d =>
-              ERN_VERSION_DIRECTORY_RE.test(d) &&
-              semver.lte(ERN_VERSION_DIRECTORY_RE.exec(d)![1], maxVersion)
-          )
-          .map(d => path.join(pathToPluginsDirectory, d))
-          .value()
-      : []
-  }
-
-  /**
-   * Given a fixed versioned plugin, return the local file system path
-   * to the directory containing the plugin configuration files, or
-   * undefined if no configuration matching this plugin version exists
-   * in the Manifest.
-   * @param plugin Fixed versioned plugin for which to retrieve configuration
-   * @param platformVersion Platform version that is querying for
-   *                        the plugin configuration.
-   */
   public async getPluginConfigurationPath(
     plugin: PackagePath,
     platformVersion: string = Platform.currentVersion
   ): Promise<string | void> {
-    if (!plugin.version) {
-      throw new Error(
-        `Plugin version is required to retrieve configuration. No version was specified for ${plugin}`
-      )
-    } else if (!semver.valid(plugin.version)) {
-      throw new Error(
-        'Plugin version is required to be fixed and not ranged to retrieve configuration'
-      )
-    }
-    let result = await this.getPluginConfigurationPathNew(
-      plugin,
-      platformVersion
-    )
-    if (!result) {
-      result = await this.getPluginConfigurationPathLegacy(
-        plugin,
-        platformVersion
-      )
-    }
-    return result
-  }
-
-  private async getPluginConfigurationPathNew(
-    plugin: PackagePath,
-    platformVersion: string = Platform.currentVersion
-  ): Promise<string | void> {
-    if (!plugin.version) {
-      throw new Error(
-        `Plugin ${PackagePath.toString()} does not have a version`
-      )
-    }
-
     await this.syncIfNeeded()
-    const versionRe = /_v(.+)\+/
-    const scopeNameRe = /^(@.+)\/(.+)$/
-    const packageNameRe = /(.+)_v\d+\.\d+\.\d+\+/
-
-    // Top level plugin configuration directories ordered by descending
-    // Electrode Native version
-    // For example :
-    // [
-    //   '/Users/blemair/.ern/ern-override-manifest/plugins/ern_v0.10.0+',
-    //   '/Users/blemair/.ern/ern-override-manifest/plugins/ern_v0.5.0+'
-    // ]
-    const orderedPluginsConfigurationDirectories = this.getPluginsConfigurationDirectories(
-      platformVersion
-    ).sort((a, b) =>
-      semver.rcompare(versionRe.exec(a)![1], versionRe.exec(b)![1])
-    )
-
-    for (const pluginsConfigurationDirectory of orderedPluginsConfigurationDirectories) {
-      let pluginScope: string | undefined
-      let pluginName: string
-      let basePluginPath: string
-      if (scopeNameRe.test(plugin.name!)) {
-        pluginScope = scopeNameRe.exec(plugin.name!)![1]
-        pluginName = scopeNameRe.exec(plugin.name!)![2]
-        basePluginPath = path.join(pluginsConfigurationDirectory, pluginScope)
-      } else {
-        pluginName = plugin.name!
-        basePluginPath = pluginsConfigurationDirectory
-      }
-
-      if (!fs.existsSync(basePluginPath)) {
-        continue
-      }
-
-      const pluginConfigDirectories = fs
-        .readdirSync(basePluginPath)
-        .filter(f => {
-          const p = packageNameRe.exec(f)
-          return p ? p![1] === pluginName : false
-        })
-
-      const pluginVersions = _.map(
-        pluginConfigDirectories,
-        s => versionRe.exec(s)![1]
-      )
-
-      const matchingVersion = _.find(pluginVersions.sort(semver.rcompare), d =>
-        semver.gte(plugin.version!, d)
-      )
-      if (matchingVersion) {
-        let pluginConfigurationPath = ''
-        pluginConfigurationPath = path.join(
-          basePluginPath,
-          `${pluginName}_v${matchingVersion}+`
-        )
-        if (
-          fs.existsSync(
-            path.join(pluginConfigurationPath, pluginConfigFileName)
-          )
-        ) {
-          return pluginConfigurationPath
-        }
-      }
-    }
-  }
-
-  /**
-   * Old way of getting plugin configuration path
-   *  -- To be deprecated --
-   */
-  private async getPluginConfigurationPathLegacy(
-    plugin: PackagePath,
-    platformVersion: string = Platform.currentVersion
-  ): Promise<string | void> {
-    if (!plugin.version) {
-      throw new Error(
-        `Plugin ${PackagePath.toString()} does not have a version`
-      )
-    }
-
-    await this.syncIfNeeded()
-    const versionRe = /_v(.+)\+/
-    const orderedPluginsConfigurationDirectories = this.getPluginsConfigurationDirectories(
-      platformVersion
-    ).sort((a, b) =>
-      semver.rcompare(versionRe.exec(a)![1], versionRe.exec(b)![1])
-    )
-
-    for (const pluginsConfigurationDirectory of orderedPluginsConfigurationDirectories) {
-      // Directory names cannot contain '/', so, replaced by ':'
-      const pluginScopeAndName = plugin.name!.replace(/\//g, ':')
-
-      const pluginVersions = _.map(
-        fs
-          .readdirSync(pluginsConfigurationDirectory)
-          .filter(f => f.startsWith(pluginScopeAndName)),
-        s => versionRe.exec(s)![1]
-      )
-
-      const matchingVersion = _.find(pluginVersions.sort(semver.rcompare), d =>
-        semver.gte(plugin.version!, d)
-      )
-      if (matchingVersion) {
-        const pluginConfigurationPath = path.join(
-          pluginsConfigurationDirectory,
-          `${pluginScopeAndName}_v${matchingVersion}+`
-        )
-        if (
-          fs.existsSync(
-            path.join(pluginConfigurationPath, pluginConfigFileName)
-          )
-        ) {
-          return pluginConfigurationPath
-        }
-      }
-    }
+    return super.getPluginConfigurationPath(plugin, platformVersion)
   }
 }
